Add explicit PETSCII/ASCII conversion for swap lists
To reduce problems with long and/or mixed-case file names in
swap lists, explicit charset conversion and an encoding marker
have been added to swap lists. By default they are assumed
to be in ASCII unless the first line in the file is "#PETSCII" (no
variation, upper-case ASCII/lower-case PETSCII characters).