fatops: Force PC64 format if the name is not FAT-compatible
Force the creation of a PC64 format file if a new file is created with
a file name that is not compatible with FAT file name rules. This
applies even when a non-PC64 file extension mode is selected, although
files with known disk-image extensions are exempt as they were before.