Optimize JiffyDOS receive timing on LPC17xx
Move the bit reception times for JiffyDOS 1.5 microseconds forward in
the LPC17xx code to avoid very rare data corruption on PAL
systems. The new timing has been tested on both PAL and NTSC. The AVR
version of the same code has not shown any errors in the same tests.