Fix handling of partition numbers in block read/write commands
Change the handling of partition numbers in block commands to be more
CMD-compatible: Instead of using the drive number in the block command
as partition, use the current partition at the time the buffer is
opened.

For additional compatibility with a few C64 titles, the partition
number in the block command is completely ignored, while the CMD
drives insist that it must be zero and throw an error otherwise.