Enable M-E handlers even if not loader was detected
The current method of deciding which handler to run for a M-E
command fails, if code wasn't transferred with a M-W command,
but for example read into a buffer from disk or transferred as
additional payload in the M-E command.

To work around this limitation, this patch changes FL_NONE to
a flag value, which can be set in the "loadertype" field of
fl_crc_table[] to indicate that this handler should be called
even if a fastloader hasn't been detected (yet).

In this case the handler is required to perform extra checks
to make sure it is responsible, e.g. by checking the checksum
of the command line or data in a buffer. If the global variable
"detected_loader" is set (that is, != FL_NONE) after the
handler returns, no other handlers will be checked or called.