Enable custom, temporary magic values for M-R handling
This enables loader handlers to set custom magic values in
response to specific events at runtime.

If the global variable "custom_magic" is != NULL, it is used if
the next M-R command references the specified address.

This has priority over the defaults defined in c1541_magics[],
so it can be used to override these.

The custom magic is reset after the next M-R command, even if it
did *not* reference the magic's address. This might not work in
cases when the result is fetched over multiple reads. But for
now be safe and maybe re-evaluate if such cases are found.