Set 0 as default result values for M-R commands
Use 0 as a defined default value for the first two bytes
returned as a result to a M-R command instead of just the
random values from error_buffer.

This makes the result predictable for addresses for which no
magic values are defined. Plus, there is no longer a need to
have entries for addresses in the c1541_magics[] table where
0 is an acceptable or useful result.