Make M-R magic values configurable for drive types
This is useful to simulate a drive corresponding to the
mounted type of D64 image (D71 => 1571, D81 => 1581).

If neither a D71 nor a D81 image is mounted (that is, a
different disk image type or none at all), the 1541 drive
is used as default.

If a custom magic has been set by a handler, this still
takes priority.