AVR: Enable longer timeouts (at reduced resolution)
The AVR code uses timer 0 for timeouts, which is an 8 bit timer.
It is configured with the /8 prescaler, so with a 8Mhz system
clock it can count up to 256 microseconds. This maximum is even
lower at higher clock speeds.

To enable longer timeouts, this patch selects the /1024 prescaler
if a timeout ist requested, which would take more than 256 timer
ticks with the /8 prescaler. This enables timeout durations of
up to 16384us with a 16MHz clock, but comes at the cost of a
reduced resolution (128us instead of 1us, with a 8MHz clock).

In addition to that, timer 2 is now used instead of timer 0. This
has the benefit that its prescaler can be reset without affecting
the other timers. Not resetting the prescaler results in a jitter
for the timeout duration of one prescaler period minus one cycle.

The if() statements should be optimized away, but compared to
the existing version, 5 more cycles are needed for the setup in
start_timeout(). This is mostly offset by the reduced jitter
from resetting the prescaler.