Enable fastloaders to request bus silence
And implement the versions used by Krill's loader in "Sonic the
Hedgehog" and "Scramble Infinity".

This enters the same sleep mode which can be selected manually
by a long press on the disk change button. Likewise, it can be
disabled again by a long press.

Disabled by default in all config files.