Enable loader-independent M-E handlers
The current method of deciding which handler to run for a M-E command
fails if code wasn't transferred with a M-W command, but for example
read into a buffer from disk or transferred as additional payload in
the M-E command.

To work around this limitation, make it possible to define a handler
function in fl_crc_table[] with loadertype FL_NONE, which will be
called if either no loader was detected or if no matching M-E handler
was found for the active loader.

Such a handler is required to perform extra checks to make sure it is
responsible, e.g. by checking the checksum of the command line or
data in a buffer.

Every handler now has to return true if it handled the call (-> loop
in run_loader() stops) or false, if not (-> search for a handler
continues).