Disable i2c if bus in invalid state
If either SCL or SDA are low in i2c_init() after configuring
both as inputs without internal pullups, this indicates a
problem with the bus. This is especially bad in case of the
SCL line, because it leads to an infinite hang in set_scl(1)
and in consequence each call to start_condition().

So abort i2c_init() and set a flag which will inhibit all i2c
communication until the next reset.