Don't close random access files at EOF
Otherwise the file has to be reopened after reading the last byte or
record, which doesn't make sense for a file used for random access.

A file is considered random access if the P (position) command has
been used on the file at least once since it was opened.

This only affects files in the FAT filesystem, as seeking in regular
(non-REL) files isn't supported for other filesystems.